# 1_python_server/swagger.yaml
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "Swagger Petstore"
  termsOfService: "http://localhost:5000"
  contact:
    email: "markus.guder@stud.uni-regensburg.de"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:5000"
basePath: "/"
schemes:
- "https"
- "http"
paths:
  /annotate/start:
    post:
      tags:
      - "annotate"
      summary: "Start annotation"
      description: ""
      operationId: "start"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Annotation Object"
        required: true
        schema:
          $ref: "#/definitions/Annotation"
      responses:
        405:
          description: "Invalid input"
  /annotate/stop:
    post:
      tags:
      - "annotate"
      summary: "Stop annotation"
      operationId: "stop"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Annotation Object"
        required: true
        schema:
          $ref: "#/definitions/Annotation"
      responses:
        405:
          description: "Invalid input"
  /annotate/next:
    post:
      tags:
      - "annotate"
      summary: "Next annotation step"
      operationId: "next"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Annotation Object"
        required: true
        schema:
          $ref: "#/definitions/Annotation"
      responses:
        405:
          description: "Invalid input"
  /meta:
    post:
      tags:
      - "annotate"
      summary: "Send meta data of annotation session"
      operationId: "meta"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Annotation Object"
        required: true
        schema:
          $ref: "#/definitions/Annotation"
      responses:
        405:
          description: "Invalid input"
  /config:
    get:
      tags:
      - "config"
      summary: "Receive configuration of clients"
      operationId: "configGET"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid status value"
    post:
      tags:
      - "config"
      summary: "Update configuration of clients session"
      operationId: "configPost"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Annotation Object"
        required: true
        schema:
          $ref: "#/definitions/Annotation"
      responses:
        405:
          description: "Invalid input"
definitions:
  Annotation:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      petId:
        type: "integer"
        format: "int64"
      quantity:
        type: "integer"
        format: "int32"
      shipDate:
        type: "string"
        format: "date-time"
      status:
        type: "string"
        description: "Order Status"
        enum:
        - "placed"
        - "approved"
        - "delivered"
      complete:
        type: "boolean"
        default: false
    xml:
      name: "Order"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"